Implemented rudiments for hierarchical approach.
